Rename gotham::http to gotham::protocol
This prevents naming conflicts of the `http` crate used in Hyper 0.12
