+++ examples/example_contribution_template/name/src/main.rs
@@ -25,10 +25,10 @@ extern crate hyper;
 extern crate mime;
 
 use hyper::{Response, StatusCode};
-use gotham::http::response::create_response;
-use gotham::state::State;
+use gotham::protocol::response::create_response;
 use gotham::router::Router;
 use gotham::router::builder::*;
+use gotham::state::State;
 
 /// Create a `Handler` that ...
 pub fn well_named_function(state: State) -> (State, Response) {
