+++ examples/handlers/async_handlers/src/main.rs
@@ -18,7 +18,7 @@ use hyper::StatusCode;
 use hyper::{Client, Uri};
 
 use gotham::handler::{HandlerFuture, IntoHandlerError};
-use gotham::http::response::create_response;
+use gotham::protocol::response::create_response;
 use gotham::router::Router;
 use gotham::router::builder::{build_simple_router, DrawRoutes};
 use gotham::router::builder::DefineSingleRoute;
