+++ examples/headers/setting/src/main.rs
@@ -6,10 +6,10 @@ extern crate hyper;
 extern crate mime;
 
 use hyper::{Response, StatusCode};
-use gotham::http::response::create_response;
-use gotham::state::State;
+use gotham::protocol::response::create_response;
 use gotham::router::Router;
 use gotham::router::builder::*;
+use gotham::state::State;
 
 // Define a custom header via the standard Hyper provided macro
 header! { (GothamHeader, "X-Gotham") => [String] }
