+++ examples/into_response/introduction/src/main.rs
@@ -10,7 +10,7 @@ extern crate serde_json;
 
 use hyper::{Response, StatusCode};
 
-use gotham::http::response::create_response;
+use gotham::protocol::response::create_response;
 use gotham::router::Router;
 use gotham::router::builder::*;
 use gotham::state::State;
