+++ gotham/src/protocol/header/x_request_id.rs
@@ -14,7 +14,7 @@ header! {
     /// # extern crate gotham;
     ///
     /// use hyper::header::Headers;
-    /// use gotham::http::header::XRequestId;
+    /// use gotham::protocol::header::XRequestId;
     ///
     /// # fn main () {
     /// let mut headers = Headers::new();
