+++ gotham_derive/src/extenders.rs
@@ -10,7 +10,7 @@ pub(crate) fn bad_request_static_response_extender(ast: &syn::DeriveInput) -> qu
             #ty_generics #where_clause
         {
             fn extend(state: &mut ::gotham::state::State, res: &mut ::hyper::Response) {
-                ::gotham::http::response::extend_response(state,
+                ::gotham::protocol::response::extend_response(state,
                                                           res,
                                                           ::hyper::StatusCode::BadRequest,
                                                           None);
